Move table methods in a different file
So that we can differentiate between scala 2 and scala 3.
