+++ src/main/scala-3/org/squeryl/internals/TypeInfo.scala
@@ -14,7 +14,6 @@ object TypeInfo {
     val uns = TypeTree.of[T]
     val symbol = uns.symbol
     val innerClassOfOptionFields: Map[String, Class[_]] = symbol.memberFields.flatMap { m =>
-      println(m)
       // we only support val fields for now
       if(m.isValDef){
         val tpe = ValDef(m, None).tpt.tpe
@@ -36,7 +35,6 @@ object TypeInfo {
       } else None
     }.toMap
 
-    println(innerClassOfOptionFields)
     Expr(innerClassOfOptionFields)
   }
